Compile Fixes.

+++ signal-service/src/main/java/com/algomeet/signalservice/dto/KyberPreKeyRequest.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class KyberPreKeyRequest {	
 	@NotNull
 	@Min(value = 1, message = "kyberPreKeyId must be greater than 0")
-	private Integer kyberPreKeyId;
+	private Long kyberPreKeyId;
 	
 	@NotEmpty
 	@Size (max = 2100)
