+++ signal-service/src/main/java/com/algomeet/signalservice/dto/KyberPreKeyResponse.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class KyberPreKeyResponse {
 	private String userKey;
 	private Integer deviceId;
-	private Integer kyberPreKeyId;
+	private Long kyberPreKeyId;
 	private String publicKey;
 	private String signature;
 	private Instant createdAt;
